requirements.txt
python-dotenv==1.2.1
groq==1.0.0
wikipedia==1.4.0
requests==2.32.5
